Return thumbnails and disc number when scanning
