+++ packages/scanner/index.d.ts
@@ -1,9 +1,10 @@
-type LocalTrack = {
+export type LocalTrack = {
     uuid: string;
-    artist?: string;
+    artist: string;
     title?: string;
     album?: string;
     duration?: number;
+    thumbnail?: Buffer;
     position?: number;
     year?: string;
 
@@ -12,5 +13,9 @@ type LocalTrack = {
     local: true;
 }
 
-declare const scanFolders = (folders: string[], supportedFormats: string[], onProgress: (progress: number, total: number, lastScanned?: string) => void) => Promise<LocalTrack[]>;
+declare const scanFolders = (
+  folders: string[], 
+  supportedFormats: string[], 
+  onProgress: (progress: number, total: number, lastScanned?: string) => void
+) => new Promise<LocalTrack[]>;
 export { scanFolders };
