+++ packages/scanner/src/lib.rs
@@ -2,61 +2,11 @@ mod error;
 mod js;
 mod local_track;
 mod scanner;
-use error::ScannerError;
-use id3::{Tag, TagLike};
-use js::{set_optional_field_str, set_optional_field_u32};
+use id3::Tag;
+use js::{set_optional_field_buffer, set_optional_field_str, set_optional_field_u32};
 use neon::prelude::*;
+use scanner::{visit_directory, visit_file};
 use std::collections::LinkedList;
-use uuid::Uuid;
-
-use local_track::LocalTrack;
-
-fn visit_file<F>(path: String, tag_reader: F) -> Result<LocalTrack, ScannerError>
-where
-    F: FnOnce(&str) -> Result<Tag, id3::Error>,
-{
-    let tag = tag_reader(&path);
-
-    match tag {
-        Ok(tag) => Ok(LocalTrack {
-            uuid: Uuid::new_v4().to_string(),
-            artist: tag.artist().map(|s| s.to_string()),
-            title: tag.title().map(|s| s.to_string()),
-            album: tag.album().map(|s| s.to_string()),
-            duration: tag.duration().unwrap_or(0),
-            position: tag.track(),
-            year: tag.year().map(|s| s as u32),
-            filename: path.split("/").last().map(|s| s.to_string()).unwrap(),
-            path: path.clone(),
-        }),
-        Err(e) => Err(ScannerError {
-            message: format!("Error reading file: {}", e),
-        }),
-    }
-}
-
-fn visit_directory(
-    path: String,
-    supported_formats: Vec<String>,
-    dirs_to_scan_queue: &mut LinkedList<String>,
-    files_to_scan_queue: &mut LinkedList<String>,
-) {
-    // Read the contents of the directory
-    let dir = std::fs::read_dir(path.clone()).unwrap();
-    for entry in dir {
-        let entry = entry.unwrap();
-        let path = entry.path();
-        if path.is_dir() {
-            // Add the directory to the queue
-            dirs_to_scan_queue.push_back(path.to_str().unwrap().to_string());
-        } else if let Some(extension) = path.extension().and_then(|ext| ext.to_str()) {
-            // Add the file to the queue, if it's a supported format
-            if supported_formats.contains(&extension.to_string()) {
-                files_to_scan_queue.push_back(path.to_str().unwrap().to_string());
-            }
-        }
-    }
-}
 
 fn scan_folders(mut cx: FunctionContext) -> JsResult<JsArray> {
     let folders: Handle<JsArray> = cx.argument(0)?;
@@ -133,16 +83,16 @@ fn scan_folders(mut cx: FunctionContext) -> JsResult<JsArray> {
         track_js_object.set(&mut cx, "uuid", track_uuid_js_string)?;
 
         set_optional_field_str(&mut cx, &mut track_js_object, "artist", track.artist);
-
         set_optional_field_str(&mut cx, &mut track_js_object, "title", track.title);
-
         set_optional_field_str(&mut cx, &mut track_js_object, "album", track.album);
 
         let track_duration_js_number = cx.number(track.duration);
         track_js_object.set(&mut cx, "duration", track_duration_js_number)?;
 
-        set_optional_field_u32(&mut cx, &mut track_js_object, "position", track.position);
+        set_optional_field_buffer(&mut cx, &mut track_js_object, "thumbnail", track.thumbnail);
 
+        set_optional_field_u32(&mut cx, &mut track_js_object, "position", track.position);
+        set_optional_field_u32(&mut cx, &mut track_js_object, "disc", track.disc);
         set_optional_field_u32(&mut cx, &mut track_js_object, "year", track.year);
 
         let track_filename_js_string = cx.string(track.filename);
@@ -174,63 +124,4 @@ fn scan_folders(mut cx: FunctionContext) -> JsResult<JsArray> {
 fn main(mut cx: ModuleContext) -> NeonResult<()> {
     cx.export_function("scanFolders", scan_folders)?;
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_sanity_check() {
-        assert_eq!(1, 1);
-    }
-
-    #[test]
-    fn test_visit_file_with_valid_file() {
-        // With mocked tag
-        let path = String::from("path/to/valid/file.mp3");
-        let result = visit_file(path.clone(), |_inner_path| {
-            let mut tag = Tag::new();
-            tag.set_artist("Artist");
-            tag.set_title("Title");
-            tag.set_album("Album");
-            tag.set_duration(123);
-            tag.set_track(1);
-            tag.set_year(2020);
-            Ok(tag)
-        });
-
-        if let Some(track) = result.ok() {
-            //check uuid format
-            assert_eq!(
-                track.uuid,
-                Uuid::parse_str(&track.uuid).unwrap().to_string()
-            );
-            assert_eq!(track.artist, Some(String::from("Artist")));
-            assert_eq!(track.title, Some(String::from("Title")));
-            assert_eq!(track.album, Some(String::from("Album")));
-            assert_eq!(track.duration, 123);
-            assert_eq!(track.position, Some(1));
-            assert_eq!(track.year, Some(2020));
-            assert_eq!(track.filename, String::from("file.mp3"));
-            assert_eq!(track.path, path);
-        } else {
-            panic!("Result is not ok");
-        }
-    }
-
-    #[test]
-    fn test_visit_file_with_no_tags() {
-        // With mocked tag
-        let path = String::from("path/to/invalid/file.mp3");
-        let result = visit_file(path.clone(), |_inner_path| {
-            Err(id3::Error::new(id3::ErrorKind::NoTag, ""))
-        });
-
-        if let Some(error) = result.err() {
-            assert_eq!(error.message, String::from("Error reading file: NoTag"));
-        } else {
-            panic!("Result is not err");
-        }
-    }
 }
