+++ packages/scanner/src/local_track.rs
@@ -5,6 +5,8 @@ pub struct LocalTrack {
     pub title: Option<String>,
     pub album: Option<String>,
     pub duration: u32,
+    pub thumbnail: Option<Vec<u8>>,
+    pub disc: Option<u32>,
     pub position: Option<u32>,
     pub year: Option<u32>,
